move Admin routes to separate file.
Also create a utility module holding the error functions.
